Fix: 수정 변수 type

See merge request s09-bigdata-recom-sub2/S09P22D206!34

+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/PowerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.pcgg.domain.recommend.entity.PowerEntity;
 
 @Repository
-public interface PowerRepository extends JpaRepository<PowerEntity, Integer> {
+public interface PowerRepository extends JpaRepository<PowerEntity, Long> {
 }
